rename top.js to top-legacy.js including test and snaps
